pmw3901 split out header and main
