+++ src/drivers/optical_flow/pmw3901/CMakeLists.txt
@@ -34,8 +34,8 @@
 px4_add_module(
 	MODULE drivers__optical_flow__pmw3901
 	MAIN pmw3901
-	COMPILE_FLAGS
-		-Wno-cast-align # TODO: fix and enable
 	SRCS
-		pmw3901.cpp
+		pmw3901_main.cpp
+		PMW3901.cpp
+		PMW3901.hpp
 	)
